taking a brake so i can learn spring boot

+++ app/src/main/java/com/larrex/purplemusic/ui/screens/AlbumDetailsScreen.kt
@@ -226,7 +226,7 @@ fun updateNowPlaying(
     musicName: String,
     artistName: String,
     albumArt: String,
-    duration: Int,
+    duration: Float,
     playingFromType: String,
     playingFromName: String,viewModel: MusicViewModel
 ) {
